Give content developer access to development tools.

+++ src/main/java/com/forgestorm/server/database/AuthenticatedUser.java
@@ -11,11 +11,11 @@ public class AuthenticatedUser {
     private final String ip;
     private final int databaseUserId;
     private final String xfAccountName;
-    private final List<Integer> secondaryGroupIds;
+    private final List<Byte> secondaryGroupIds;
     private final boolean isAdmin;
     private final boolean isModerator;
 
-    public AuthenticatedUser(final String ip, final int databaseUserId, final String xfAccountName, final List<Integer> secondaryGroupIds, final boolean isAdmin, final boolean isModerator) {
+    public AuthenticatedUser(final String ip, final int databaseUserId, final String xfAccountName, final List<Byte> secondaryGroupIds, final boolean isAdmin, final boolean isModerator) {
         this.ip = ip;
         this.databaseUserId = databaseUserId;
         this.xfAccountName = xfAccountName;
