+++ src/main/java/com/forgestorm/server/network/login/LoginState.java
@@ -11,7 +11,7 @@ public class LoginState {
     private Boolean loginSuccess;
     private LoginFailReason loginFailReason;
     private String username;
-    private List<Integer> secondaryGroupIds;
+    private List<Byte> secondaryGroupIds;
     private boolean isAdmin;
     private boolean isModerator;
 
@@ -21,7 +21,7 @@ public class LoginState {
         return this;
     }
 
-    public LoginState successState(int userId, String username, List<Integer> secondaryGroupIds, boolean isAdmin, boolean isModerator) {
+    public LoginState successState(int userId, String username, List<Byte> secondaryGroupIds, boolean isAdmin, boolean isModerator) {
         this.userId = userId;
         loginSuccess = true;
         this.username = username;
